third_party/blink/web_tests/accessibility/aria-fallback-roles-expected.txt

test1 test2 test3 test4

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test.role is expectedRole
Role should be: AXRole: AXCheckBox
PASS test.role is expectedRole
Role should be: AXRole: AXButton
PASS test.role is expectedRole
Role should be: AXRole: AXButton
PASS test.role is expectedRole
Role should be: AXRole: AXButton
PASS successfullyParsed is true

TEST COMPLETE
